+++ test/unittest/src/sm4_common.cpp
@@ -116,7 +116,7 @@ int32_t Sm4EncryptWithInput(HcfCipher *cipher, HcfSymKey *key, HcfBlob *input,
 int32_t Sm4DecryptEmptyMsg(HcfCipher *cipher, HcfSymKey *key, HcfParamsSpec *params,
     uint8_t *cipherText, int cipherTextLen)
 {
-    HcfBlob input = { .data = cipherText, .len = cipherTextLen };
+    HcfBlob input = { .data = cipherText, .len = static_cast<size_t>(cipherTextLen) };
     HcfBlob output = { .data = nullptr, .len = 0 };
     int32_t ret = cipher->init(cipher, DECRYPT_MODE, &(key->key), params);
     if (ret != 0) {
@@ -186,7 +186,7 @@ int32_t Sm4Decrypt(HcfCipher *cipher, HcfSymKey *key, HcfParamsSpec *params,
     uint8_t *cipherText, int cipherTextLen)
 {
     uint8_t plainText[] = "this is test!";
-    HcfBlob input = {.data = reinterpret_cast<uint8_t *>(cipherText), .len = cipherTextLen};
+    HcfBlob input = {.data = reinterpret_cast<uint8_t *>(cipherText), .len = static_cast<size_t>(cipherTextLen)};
     HcfBlob output = {};
     int32_t maxLen = cipherTextLen;
     int32_t ret = cipher->init(cipher, DECRYPT_MODE, reinterpret_cast<HcfKey *>(key), params);
@@ -263,7 +263,7 @@ int32_t Sm4NoUpdateDecrypt(HcfCipher *cipher, HcfSymKey *key, HcfParamsSpec *par
     uint8_t *cipherText, int cipherTextLen)
 {
     uint8_t plainText[] = "this is test!";
-    HcfBlob input = {.data = reinterpret_cast<uint8_t *>(cipherText), .len = cipherTextLen};
+    HcfBlob input = {.data = reinterpret_cast<uint8_t *>(cipherText), .len = static_cast<size_t>(cipherTextLen)};
     HcfBlob output = {};
     int32_t maxLen = cipherTextLen;
     int32_t ret = cipher->init(cipher, DECRYPT_MODE, reinterpret_cast<HcfKey *>(key), params);
